# public/swagger/api.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a api document of the sample node.js server framework.",
        "version": "1.0.0",
        "title": "RESTful API Document",
        "termsOfService": "",
        "contact": {
            "email": "super_xp@126.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "host": "",
    "basePath": "/",
...
    "schemes": [
        "http"
    ],
    "paths": {
        "/api/auth/signInWithEmailAndPassword": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "用邮箱地址和密码登录",
                "description": "用邮箱地址和密码登录",
                "operationId": "signInWithEmailAndPassword",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Created user object",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/SingUp_Email_Data"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        },
                        "headers": {
                            "X-Rate-Limit": {
                                "type": "integer",
                                "format": "int32",
                                "description": "calls per hour allowed by the user"
                            },
                            "X-Expires-After": {
                                "type": "string",
                                "format": "date-time",
                                "description": "date in UTC when token expires"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid username/password supplied"
                    }
                }
            }
        },
...
        "/api/auth/sendPasswordResetEmail": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "通过邮箱重置密码",
                "description": "",
                "operationId": "sendPasswordResetEmail",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Created user object",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/SendPasswordResetEmail_Email_Data"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
        "/api/auth/resendStringTokenByEmail": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "重新向邮箱发送stringtoken",
                "description": "重新向邮箱发送stringtoken",
                "operationId": "resendStringTokenByEmail",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Created user object",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/resendStringTokenByEmail_Data"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
        "/api/utils/captcha": {
            "get": {
                "tags": [
                    "utils"
                ],
                "summary": "获取验证码",
                "description": "获取验证码",
                "operationId": "captcha",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "width",
                        "in": "query",
                        "description": "宽",
                        "required": true,
                        "type": "string",
                        "default": "80"
                    },
                    {
                        "name": "height",
                        "in": "query",
                        "description": "高",
                        "required": true,
                        "type": "string",
                        "default": "30"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        },
                        "headers": {
                            "X-Rate-Limit": {
                                "type": "integer",
                                "format": "int32",
                                "description": "calls per hour allowed by the user"
                            },
                            "X-Expires-After": {
                                "type": "string",
                                "format": "date-time",
                                "description": "date in UTC when token expires"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid username/password supplied"
                    }
                }
            }
        },
...
        "/data/data/findandupdate": {
            "post": {
                "tags": [
                    "data"
                ],
                "summary": "数据更新",
                "description": "数据更新",
                "operationId": "findandupdate",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Created user object",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/test_data_update"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/data/data/delete": {
            "post": {
                "tags": [
                    "data"
                ],
                "summary": "数据更新",
                "description": "数据更新",
                "operationId": "findandupdate",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Created user object",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/test_data_delete"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "api_key": {
            "type": "apiKey",
            "name": "authorization",
            "in": "header"
        }
    },
    "definitions": {
...
        "SingUp_Email_Data": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "111@111.com"
                },
                "password": {
                    "type": "string",
                    "example": "111111"
                }
            },
            "xml": {
                "name": "SingUp_Email_Data"
            }
        },
        "SendPasswordResetEmail_Email_Data": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "111@111.com"
                }
            },
            "xml": {
                "name": "SendPasswordResetEmail_Email_Data"
            }
        },
        "resendStringTokenByEmail_Data": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "111@111.com"
                },
                "type": {
                    "type": "string",
                    "example": "sign/resetpass"
                }
            },
            "xml": {
                "name": "SendPasswordResetEmail_Email_Data"
            }
        },
...
        "test_data_update": {
            "type": "object",
            "properties": {
                "document": {
                    "type": "string",
                    "example": "testabc"
                },
                "operation": {
                    "type": "string",
                    "example": "update"
                },
                "find": {
                    "$ref": "#/definitions/test_document_find"
                },
                "update": {
                    "$ref": "#/definitions/test_document_update"
                }
            },
            "xml": {
                "name": "test_data_update"
            }
        },
        "test_document_find": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "111@111.com"
                },
                "auth": {
                    "type": "object",
                    "properties": {
                        "role": {
                            "type": "boolean",
                            "example": false
                        },
                        "testaa": {
                            "type": "object",
                            "properties": {
                                "bb": {
                                    "type": "string",
                                    "example": "321321"
                                }
                            }
                        }
                    }
                }
            },
            "xml": {
                "name": "test_document_find"
            }
        },
        "test_document_update": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "testname11"
                }
            },
            "xml": {
                "name": "test_document_update"
            }
        },
        "test_data_delete": {
            "type": "object",
            "properties": {
                "document": {
                    "type": "string",
                    "example": "testabc"
                },
                "operation": {
                    "type": "string",
                    "example": "delete"
                },
                "find": {
                    "$ref": "#/definitions/test_document_delete_find"
                }
            },
            "xml": {
                "name": "test_data_delete"
            }
        },
        "test_document_delete_find": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "testname12"
                },
                "auth": {
                    "type": "object",
                    "properties": {
                        "testaa": {
                            "type": "object",
                            "properties": {
                                "bb": {
                                    "type": "object",
                                    "properties": {
                                        "$lt": {
                                            "type": "string",
                                            "example": "321321"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "xml": {
                "name": "test_document_delete_find"
            }
        },
...
    },
...
}